[Var] ancient: Update to v2.1.0 (2023-07-11).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@19455 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ include/ancient/src/HUFFDecompressor.cpp
@@ -60,7 +60,7 @@ void HUFFDecompressor::decompressImpl(Buffer &rawData,const Buffer &previousData
 
 	ForwardOutputStream outputStream(rawData,0,rawData.size());
 
-	HuffmanDecoder<uint32_t> decoder;
+	HuffmanDecoder<uint8_t> decoder;
 	for (uint32_t i=0;i<256;i++)
 	{
 		uint8_t codeBits=readByte()+1;
@@ -74,7 +74,7 @@ void HUFFDecompressor::decompressImpl(Buffer &rawData,const Buffer &previousData
 			shift+=8;
 		}
 		code=(code>>shift)&((1<<codeBits)-1);
-		decoder.insert(HuffmanCode<uint32_t>{codeBits,code,i});
+		decoder.insert(HuffmanCode<uint8_t>{codeBits,code,uint8_t(i)});
 	}
 
 	while (!outputStream.eof())
